fix(workflow): add release creation step to maven-central.yml (development)
Add a new step in the GitHub Actions workflow for creating a release
after a successful deployment. This step uses the comnoco/create-release
action to automate the creation of a release with the specified tag
name and release name. It also sets the draft and prerelease status
based on the release version.